pyproject.toml
[tool.poetry]
name = "LLMToolkit"
version = "0.1.0"
description = "An open-source project providing a unified interface to integrate Large Language Models (LLMs) with various services."
authors = ["Marcos Ventosa"]
license = "MIT"
readme = "README.md"

# Include the 'llmtoolkit' package from the 'src' directory
packages = [
    { include = "llmtoolkit", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
tiktoken = "^0.8.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D"  # pydocstyle
]
ignore = [
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "D401"  # imperative mood
]

[tool.ruff.lint.pydocstyle]
convention = "google"
